:bricks: move header data type, header data fetcher into their own files
