+++ next/components/header/header.tsx
@@ -1,41 +1,8 @@
-import { headerSchema } from "@/schemas/single/Header";
 import { getLargestImageUrl } from "@/util/image";
-import fetchStrapi from "@/util/strapi";
 import Image from "next/image";
 import Link from "next/link";
-import { z } from "zod";
 import styles from "./header.module.css";
-
-async function getHeaderData() {
-  const resData = await fetchStrapi("header", headerSchema);
-  const membersData = await fetchStrapi("member-teams", z.any());
-
-  // Ensure membersData is an array and sort in descending order
-  const members = Array.isArray(membersData)
-    ? membersData
-        .map((item: any) => ({
-          ...item,
-          CommitteeYear: Number(item.CommitteeYear), // Convert to number for sorting
-        }))
-        .sort((a, b) => b.CommitteeYear - a.CommitteeYear) // Sort in descending order
-    : []; // Default to an empty array if not an array
-
-  const projects = [
-    {
-      Title: new Date().getFullYear(),
-      slug: "current",
-    },
-    {
-      Title: "Past",
-      slug: "past",
-    },
-  ];
-  return {
-    ...resData,
-    members,
-    projects,
-  };
-}
+import { getHeaderData } from "./headerDataFetcher"; // Adjust the path as necessary
 
 export default async function Header() {
   const data = await getHeaderData();
